not a lot really.
Still trying to figure out the best way to do the delete. Obviously it
has to use a form submit so that it's a POST otherwise it can't really
be secure. By I want to confirm the delete so I can't just embed a form
in each row. Actually I could. I'd just have to pass the id of the form
to submit to the confirm script somehow. Bitty.