Finished adding and removing friends to an existing event.
This is all wrong at the moment because you can only add friends to an
event that has already been created. It needs to work client side I
think so that we can save the friends when we save the event. This might
be a good chance to try out angular. Or just do it manually which will
not be very hard.