Continued the refactor towards angularisation
Converted the events list to the angular way.
need to work out how to load an existing event. Need to pick the id out
of the route but I'm not using the angular router.
Delete doesn't work at the moment.
Converted all my json responses to use the JSON.Net library so it
handles dates properly.